Add packaged CLI entry point
