+++ nhl_commentary_core/__main__.py
@@ -1,0 +1,6 @@
+"""Module entry point for ``python -m nhl_commentary_core``."""
+
+from .cli import main
+
+if __name__ == "__main__":  # pragma: no cover - exercised via manual invocation
+    main()
